+++ packages/meteor-canvas/lib/CanvasComponent.js
@@ -122,6 +122,29 @@ CanvasComponent = class {
         y: this.y + y
       });
     } else {
+      return { x: this.x + x, y: this.y + y };
+    }
+  }
+
+  // globalToLocal({x, y})
+  // globalToLocal(x, y)
+  globalToLocal(x, y) {
+    if (x && typeof x === 'object') {
+      y = x.y;
+      x = x.x;
+    }
+    // NOTE: We only permit translation operations
+    // so we just need to concern ourselves with x and y.
+    // If we were to permit scaling and rotating components
+    // then we would need each component maintain a local
+    // matrix transform and we would have to compose matrices
+    // up the entire hierarchy.
+    if (this.parent) {
+      let p = this.parent.globalToLocal(x, y);
+      p.x -= this.x;
+      p.y -= this.y;
+      return p;
+    } else {
       return { x: x, y: y };
     }
   }